Switching to physics
This was basically just switching from moving the motors on a scale of motor power, to looking at the PID values from the wheel and switching to feet/sec. Hopefully this makes autonomous run smoother so they will all go the exact same speed and the calculations won't be off.
right now the program is set up to use the smartdashboard to get the values for the kP and kFF really easily so you can just get them and plug them straight in for your robot :)