Mostly working
So right now it's a little bit of a mess, everything works with commands like the DriveCommand that lets you drive from where you are to another point and there is also Obstacle command that lets you drive around an object. We are working on making it drive a path. We also got LEDs working.